refactor(workflow-tasks): move cleanup component into library
